xong bài 110, nhớ add thêm cột CapitalPrice ở bảng Products trong sql

+++ ECommerce_web/Models/StatisticalModel.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int Quantity { get; set; } //So luong ban
         public int Sold { get; set; } //So lượng đơn hàng
-        public int Revenue { get; set; } //Doanh thu
-        public int Profit { get; set; } //Lợi nhuận
+        public decimal Revenue { get; set; } //Doanh thu
+        public decimal Profit { get; set; } //Lợi nhuận
         public DateTime DateCreated { get; set; } //Ngày bán
     }
 }
